third_party/WebKit/LayoutTests/fast/dom/DeviceMotion/window-property-expected.txt

Tests that the window.DeviceMotionEvent and window.ondevicemotion properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL typeof window.DeviceMotionEvent == 'object' should be true. Was false.
PASS typeof window.DeviceMotionEvent == 'function' is false
FAIL hasDeviceMotionEventProperty() should be true. Was false.
FAIL 'DeviceMotionEvent' in window should be true. Was false.
FAIL window.hasOwnProperty('DeviceMotionEvent') should be true. Was false.
FAIL typeof window.ondevicemotion == 'object' should be true. Was false.
FAIL hasOnDeviceMotionProperty() should be true. Was false.
FAIL 'ondevicemotion' in window should be true. Was false.
FAIL window.hasOwnProperty('ondevicemotion') should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
